requirements.txt
google-cloud-speech>=2.35.0
google-cloud-storage>=2.18.0
sounddevice>=0.5.3
pydub>=0.25.1
numpy>=2.4.0
scipy>=1.16.3
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart>=0.0.9
